requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
duckduckgo_search==8.0.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
ftfy==6.3.1
h11==0.14.0
hnswlib==0.8.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
open_clip_torch==2.32.0
packaging==24.2
pillow==11.2.1
primp==0.15.0
pydantic==2.11.3
pydantic_core==2.33.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
timm==1.0.15
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
wcwidth==0.2.13
